+++ CHANGELOG.md
@@ -23,11 +23,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make stacked bar chart a subcomponent of cartesian chart.
 - Make point chart a subcomponent of cartesian chart.
 - Use flipped property of cartesian charts to flip bar charts.
+- Rename DataBarsCreation interface to DataSeriesBarCreation.
+- Rename DataBarsGroupedCreation interface to DataSeriesBarGroupedCreation.
+- Rename DataBarsStackedCreation interface to DataSeriesBarStackedCreation.
+- Rename DataPointsCreation interface to DataSeriesPointCreation.
 
 ### Removed
 
 - Remove bar orientation enum.
-- Remove orientation property from bar series and bar charts data. 
+- Remove orientation property from bar series and bar charts data.
+- Remove DataSeriesBarCustom interface & dataSeriesBarCustom function.
+- Remove DataSeriesPointCustom interface & dataSeriesPointCustom function.
+- Remove creation property from bar series data.
+- Remove creation property from grouped bar series data.
+- Remove creation property from stacked bar series data.
+- Remove creation property from point series data.
 
 ### Fixed